moved some ecnf upload functions from import_utils back to import_ecnf_data and fixed galrep upload
